requirements.txt
torch
torchvision
opecv-python
pillow
pyYAML
tqdm
matplotlib
seaborn
pandas
flask
requests
scipy
lime